完成修改密码，细节修复

+++ TojoMobile/user/businessModel/TJUserBasicInfoModel.h
@@ -17,7 +17,7 @@
 @property (nonatomic, assign) int userId;
 @property (nonatomic, copy) NSString *userEmail;
 @property (nonatomic, copy) NSString *userRealName;
-@property (nonatomic, assign) short userGender;
+@property (nonatomic, assign) int userGender;
 @property (nonatomic, copy) NSString *userUniversity;
 @property (nonatomic, copy) NSString *userImage;
